+++ src/app/robots.ts
@@ -1,0 +1,18 @@
+import { MetadataRoute } from "next";
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: {
+      userAgent: "*",
+      allow: ["/", "/order/", "/menu/"],
+      disallow: [
+        "/thisisforadmin/",
+        "/checkout/",
+        "/order-complete/",
+        "/payment/",
+        "/api/",
+      ],
+    },
+    sitemap: "https://easysalad.co.kr/sitemap.xml",
+  };
+}
